Adding gitprompt + git aliases/removing stray file
Adding the gitprompt() function so I can turn on and off the git portion
of my prompt at will.

Changing the status icon to show something different for not having git
installed, having the prompt disabled, or just being in a directory
that's not a git repo.

Removing an old file that just had errors from a debugging session.