Adding the ability to sync without rsync
Right.  I have now found an environment in which rsync is *NOT*
installed by default.  The mind boggles.

Splitting out the standard rsync command into a wrapping function, and
adding the equivalent tar invocation in its own wrapper.

Adding detection - if rsync is there, I use that; if not, it's tar.

NOTE: I'm not 100% sure how portable the tar '--exclude-from' syntax is.
Something to watch for.

Correcting the rsync inclusion/exclusion arguments.  Didn't work the way
I thought it did.