Abandoning regex branch determination for now
No amount of expansion of the original $IFS, nor other massaging of the
character class, seemed to give the branch name in BASH_REMATCH.  It
either ended up blank or containing some portion of the next line, with
the newline included.

Switching to using awk for branch determination in the prompt for the
time being.